backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiortc>=1.13.0",
    "fastapi[standard]>=0.116.1",
    "joblib>=1.5.2",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "uvicorn[standard]>=0.30.0",
    "av>=12.0.0",
    "websockets>=12.0",
    "requests>=2.32.5",
    "facenet-pytorch>=2.5.3",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "cryptography>=45.0.7",
    "sqlalchemy>=2.0.43",
    "pgvector>=0.4.1",
    "faiss-cpu>=1.12.0",
    "psycopg[binary,pool]>=3.2.10",
    "requests-async>=0.2.4",
]
